restructure theme folder
