+++ App.js
@@ -1,11 +1,11 @@
 import { StatusBar } from "expo-status-bar";
 import AppNavigation from "./src/navigation";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ThemeProvider as ThemeProviderV2 } from "@/hooks/theme/ThemeContext";
 import { ThemeProvider as ThemeProviderV1 } from "./src/utils/Context";
-import { useThemeColor } from "@/hooks/theme/useThemeColor";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { useThemeColor } from "@/theme/useThemeColor";
+import { ThemeProvider as ThemeProviderV2 } from "@/theme/ThemeContext";
 
 const queryClient = new QueryClient();
