+++ src/components/themed/ThemedTextInput.tsx
@@ -1,4 +1,4 @@
-import { useThemeColor } from "@/hooks/theme/useThemeColor";
+import { useThemeColor } from "@/theme/useThemeColor";
 import { TextInput, TextInputProps } from "react-native";
 
 export type ThemedTextInputProps = TextInputProps & {
