+++ src/theme/ThemeContext.js
@@ -1,11 +1,11 @@
 import React, { createContext, useCallback, useContext } from "react";
 import { useColorScheme } from "./useColorScheme";
-import { Colors } from "@/constants/Colors";
 import {
 	DarkTheme,
 	DefaultTheme,
 	ThemeProvider as ThemeProviderDefault,
 } from "@react-navigation/native";
+import { Colors } from "./Colors";
 
 const ThemeContext = createContext();
